==> python code/01 - Dataset_collection.ipynb <==
# Dataset Collection


The Dataset is collected from HuggingFace, which contains approximately 4.5M entries in it. It has the data split into - Categories, Countries and Currencies. 

## Dataset Structure

Each record has the following fields:
```json
{
  "transaction_description": "string",
  "category": "string", 
  "country": "string",
  "currency": "string"
}
```

## Example records:

| transaction_description | category                      | country    | currency |
|--------------------------|-------------------------------|------------|----------|
| McDonald's #1234         | Food & Dining                | USA        | USD      |
| Uber Ride                | Transportation               | UK         | GBP      |
| Amazon Purchase          | Shopping & Retail            | CANADA     | CAD      |
| Netflix Subscription     | Entertainment & Recreation   | AUSTRALIA  | AUD      |
| Pharmacy Purchase        | Healthcare & Medical         | INDIA      | INR      |

# Installing Libraries
Before proceeding with any work, we should install the required python libraries
```bash
pip install datasets
pip install pandas
```

# Import Packages
Import nessasary packages into the notebook

from datasets import load_dataset
import pandas as pd

# Download Dataset
Download the required dataset from HuggingFace

dataset = load_dataset("mitulshah/transaction-categorization")

print(dataset)

dataset['train'].to_csv("datasets/Dataset.csv")

dataset['train']

# Dataset Scaling
Working with 4.5M entries is resource intensive, so we scale down the dataset to just have only 200k entries

small_dataset = dataset['train'].shuffle(seed=42).select(range(400000))
small_dataset.to_csv("datasets/Dataset_small.csv")

# Dataset Check
Checking the scaled-down Dataset would help prevent issues later on.
(Better to be safe than sorry!)

df = pd.read_csv("datasets/Dataset_small.csv")

df.head()

df.shape
