==> python code/03 - Dataset_cleaning.ipynb <==
# Import Packages

Before starting the cleaning process, import all required packages

import pandas as pd
import re

# Import the Dataset

Import the Analyzed Dataset to process it

df = pd.read_csv("datasets/Dataset_small.csv")

# Create Cleaner Function
Create a function that selectively removes all unwanted Strings like `TXN363507` or `#7300`. The function developed does the following work:
- Converts the Text to lowercase
- Removes predictable patterns
- Removes stray Numbers and Special Characters
- Returns the text

def clean_text(text):
    # Lowercase text
    text = text.lower()
    
    # Remove TXN patterns
    text = re.sub(r'txn\d+', '', text)

    # Remove hashtags patterns
    text = re.sub(r'#\d+', '', text)
    
    # Remove standalone numbers
    text = re.sub(r'\d+', '', text)
    
    # Remove special characters
    text = re.sub(r'[^a-z\s]', '', text)
    
    # Remove extra spaces
    text = re.sub(r'\s+', ' ', text).strip()
    
    return text

# Testing the function

test_samples = [
    "Mobile Center TXN797664",
    "Hulu #5061 TXN363507",
    "PNC Bank - INDIA (Digital Wallet)",
    "Zaxby's #3457 (Digital Wallet)"
]

for t in test_samples:
    print("Original:", t)
    print("Cleaned :", clean_text(t))
    print()

# Perform Cleaning

df['transaction_description_c'] = df['transaction_description'].apply(clean_text)

df[['transaction_description', 'transaction_description_c', 'category']].head(10)

# Verify Cleaning

# Check for existance of Numbers
df['transaction_description_c'].str.contains(r'\d').sum()

# Check for existance of Special Characters
df['transaction_description_c'].str.contains(r'[^a-z\s]').sum()

# Visually compare outputs of the Cleaning
df[['transaction_description', 'transaction_description_c']].sample(20)

df = df.drop(columns=['transaction_description'])

df.to_csv('datasets/Dataset_cleaned.csv', index=False)
