==> python code/04 - Feature_Engineering_and_Model_Training.ipynb <==
# Install Libraries

Before starting, install all required Libraries onto the System by using the following commands:
```bash
pip install scikit-learn
pip install pandas
```

# Import Packages

Import all required packages into the notebook

from sklearn.metrics import classification_report, accuracy_score
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LogisticRegression
from sklearn.pipeline import FeatureUnion
import pandas as pd
import pickle
import re
import os

# Load Dataset

Now load the cleaned dataset into the notebook

df = pd.read_csv("datasets/Dataset_cleaned.csv")


df.head()

df.shape

# Train-Test Splitting

Split the dataset into `test` and `train` datasets for the model. We use `stratify = y` for making sure that the classes are balanced in train-test splits

X = df['transaction_description_c']
y = df['category']

X_train, X_test, y_train, y_test = train_test_split(
    X,
    y,
    test_size = 0.2,
    random_state = 42,
    stratify = y
)

# TF-IDF Vectorization

TF-IDF (Term Frequency - Inverse Document Frequency). This is one of the most important step for the NLP, which converts text into numerical features, which can then be used further by the model training. Here, we use a **Unigram** and a **Bigram** which works fine with short-text features. 

We are limiting the `max_features` to `1000`, just t ensure the model takes in the most important tokens, making the model lightweight.

We are also limiting `min_df` to `2`, which ensures that the model takes in the words which appear in atleast `2` documents, which avoids taking in extremely rare tokens as important tokens

# Defining the Word Vectorizer behaviour
word_vectorizer = TfidfVectorizer(
    analyzer='word',
    ngram_range=(1,2),
    max_features=10000,
    min_df=2
)

# Defining the Character Vectorizer behaviour
char_vectorizer = TfidfVectorizer(
    analyzer='char_wb',
    ngram_range=(3,5),
    max_features=10000,
    min_df=2
)

# Combined Vectorizers
vectorizer = FeatureUnion([
    ("word_tfidf", word_vectorizer),
    ("char_tfidf", char_vectorizer)
])

# Applying Vectorizer
X_train = vectorizer.fit_transform(X_train)
X_test = vectorizer.transform(X_test)

# Checking Vectorized shapes
print("Train shape:", X_train.shape)
print("Test shape:", X_test.shape)

# Model Training

We used **Logistic Regression** as our classification model, trained on top of **TF-IDF vectorized text features**. Logistic Regression is a strong baseline for text classification tasks. When combined with TF-IDF features, it performs efficiently and provides reliable results.


model = LogisticRegression(max_iter=2000)

model.fit(X_train, y_train)

model.n_iter_

# Model Evaluation

In this stage, we evaluate our model's perfomance by analyzing it's accuracy and manually testing it's outputs

y_pred = model.predict(X_test)

print("Accuracy:", accuracy_score(y_test, y_pred))
print(classification_report(y_test, y_pred))

## Manual Testing

def clean_text(text):
    # Lowercase text
    text = text.lower()
    
    # Remove TXN patterns
    text = re.sub(r'txn\d+', '', text)

    # Remove hashtags patterns
    text = re.sub(r'#\d+', '', text)
    
    # Remove standalone numbers
    text = re.sub(r'\d+', '', text)
    
    # Remove special characters
    text = re.sub(r'[^a-z\s]', '', text)
    
    # Remove extra spaces
    text = re.sub(r'\s+', ' ', text).strip()
    
    return text

def predict_category(text):
    # Clean input text
    cleaned = clean_text(text)
    
    # Convert to vector using trained vectorizer
    vector = vectorizer.transform([cleaned])
    
    # Predict category
    prediction = model.predict(vector)[0]
    
    return prediction

predict_category("Amazon Order")

predict_category("Uber Ride")

predict_category("Tacobell")

predict_category("Taco bell")

predict_category("Chicken Pot Pie")

predict_category("Lotus Resorts")

# Exporting Model

Since we have trained the model using a Vectorizer and SVM, we can proceed to save the model so that it can be used further

# Create models folder if not exists
os.makedirs("models", exist_ok=True)

# Save model
with open("models/model.pkl", "wb") as f:
    pickle.dump(model, f)

# Save vectorizer
with open("models/vectorizer.pkl", "wb") as f:
    pickle.dump(vectorizer, f)

# Test Loading

We need to make sure that our exported model is stable enough to be portable, so we load back the exported model and test it

with open("models/model.pkl", "rb") as f:
    loaded_model = pickle.load(f)

with open("models/vectorizer.pkl", "rb") as f:
    loaded_vectorizer = pickle.load(f)

def predict_loaded(text):
    cleaned = clean_text(text)
    vector = loaded_vectorizer.transform([cleaned])
    return loaded_model.predict(vector)[0]

predict_loaded("tacobell")

predict_loaded("Amazon Delivery")
